fileroller: extended extraction functions
The main addition of this commit is separate the "Extract Here" functionality
into two separate functions. Traditionally, would use file-roller's
"extract-here" flag which will try to extract into a sub-folder every time,
creating one automatically if an encompassing folder doesn't exist in the
archive root.

This commit makes the action more explicit, either extracting "here" as in
in this folder, or extracting to a sub-folder, in which case a sub-folder will
always be created.

- Enables "Extract To..." menu item
- Adds "Extract To Sub-Folder" function.
- Makes "Extract Here" always extract files into the archive's folder.